feat: Add 404 error handlers for bootcamps

+++ backend/controllers/bootcamp.py
@@ -2,49 +2,15 @@ from flask import jsonify
 from models.bootcamp import Bootcamp
 
 
-'''
-    GET /api/v1/bootcamps
-        Returns status code 200 and json object { "success": True, "data": bootcamps}
-            where bootcamps is the list of all bootcamps
-        Access public
-'''
-
-
 def get_bootcamps():
-    bootcamps = Bootcamp.query.all()
-    bootcamps = [bootcamp.format() for bootcamp in bootcamps]
-    data = jsonify({
-        "success": True,
-        "data": bootcamps
-    })
-    return data
-
-
-'''
-    GET /api/v1/bootcamps/<int:id>
-        Returns status code 200 and json object { "success": True, "data": bootcamp}
-            where bootcamp is the bootcamp with the id of id
-            that is defined within the query string
-        Access public
-'''
+    return Bootcamp.query.all()
 
 
 def get_single_bootcamp(id):
-    bootcamp = Bootcamp.query.filter_by(id=id).one_or_none()
-    data = jsonify({
-        "success": True,
-        "data": bootcamp.format()
-    })
-
-    return data
-
-
-'''
-    POST /api/v1/bootcamps
-        Returns status code 201 and json object { "success": True, "data": bootcamp}
-            where bootcamp is the newly create bootcamp
-        Access Private
-'''
+    try:
+        return Bootcamp.query.filter_by(id=id).one_or_none()
+    except:
+        return None
 
 
 def add_bootcamp(request):
@@ -64,25 +30,14 @@ def add_bootcamp(request):
 
     new_bootcamp.insert()
 
-    data = jsonify({
-        "success": True,
-        "message": new_bootcamp.format()
-    })
-
-    return data
-
-
-'''
-    PUT /api/v1/bootcamps/<int:id>
-        Returns status code 200 and json object { "success": True, "data": bootcamp}
-            where bootcamp is the updated bootcamp with the id of id
-            that is defined within the query string
-        Access private
-'''
+    return new_bootcamp
 
 
 def update_bootcamp(request, id):
     bootcamp = Bootcamp.query.filter_by(id=id).one_or_none()
+
+    if bootcamp is None:
+        return None
 
     updated_bootcamp = request.get_json()
 
@@ -97,23 +52,15 @@ def update_bootcamp(request, id):
 
     bootcamp.update()
 
-    data = jsonify({
-        "success": True,
-        "message": bootcamp.format()
-    })
-    return data
-
-
-'''
-    DELETE /api/v1/bootcamps/<int:id>
-        Returns status code 200 and json object { "success": True }
-            
-        Access private
-'''
+    return bootcamp
 
 
 def delete_bootcamp(id):
     bootcamp = Bootcamp.query.filter_by(id=id).one_or_none()
+
+    if bootcamp is None:
+        return None
+
     bootcamp.delete()
     data = jsonify({
         "success": True
